human_match/requirements.txt
Django==3.0.5
djangorestframework==3.11.0
mysqlclient==1.4.6
pillow==7.1.1
pytest-django==3.9.0